test: split shell reducer contract suite
